Refactor : fixed few warnings and moved data fetching to utils
