+++ src/components/Time.jsx
@@ -23,14 +23,16 @@ const Time = () => {
         secs: date.getSeconds(),
       },
     });
-  }, []);
+  }, [dataDispatch]);
 
   const { hours, mins, secs } = time;
 
   return (
-    <h1 className="txt-5xl txt-bold">{`${getFormattedvalue(
-      hours
-    )} : ${getFormattedvalue(mins)} : ${getFormattedvalue(secs)}`}</h1>
+    <h1 className="txt-5xl txt-bold">
+      {`${getFormattedvalue(hours)} : ${getFormattedvalue(
+        mins
+      )} : ${getFormattedvalue(secs)}`}{" "}
+    </h1>
   );
 };
