+++ src/utils/index.js
@@ -1,3 +1,7 @@
 export { initialDataState, DATA_ACTIONS } from "./constants";
 
 export { quotesDB } from "./quotesDB";
+
+export { getWeatherData } from "./getWeatherData";
+
+export { getBackgroundImage } from "./getBackgroundImage";
